requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
deepseek-ai==0.0.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
langchain==0.1.0
numpy==1.24.3 